requirements.txt
fastapi==0.103.2
uvicorn==0.23.2
requests==2.31.0
pydantic==2.4.2
python-dotenv==1.0.0